requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# HTTP and A2A Protocol dependencies
httpx>=0.25.0
agentbeats>=0.1.0
a2a>=0.1.0

# OpenAI (Required for Phase 1 & 2)
openai>=1.0.0

# Anthropic Claude (Optional - Phase 2 only)
anthropic>=0.18.0

# Google Gemini (Optional - Phase 2 only)
google-generativeai>=0.3.0

# Optional: Enhanced logging
colorama>=0.4.6

# Optional: Progress bars
tqdm>=4.65.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
